requirements.txt
flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.0
eventlet==0.36.1
psycopg2-binary==2.9.9
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
PyPDF2==3.0.1
python-docx==1.1.2
pyarrow==16.1.0
fastparquet==2024.5.0
openpyxl==3.1.4
python-dotenv==1.0.1
gunicorn==22.0.0
